combine GeneralStream/BatchTableEnvironment to Stream/BatchTableEnvironment
