+++ flink-connectors/flink-orc/src/test/java/org/apache/flink/orc/OrcTableSourceITCase.java
@@ -20,7 +20,7 @@ package org.apache.flink.orc;
 
 import org.apache.flink.api.java.DataSet;
 import org.apache.flink.api.java.ExecutionEnvironment;
-import org.apache.flink.table.api.GeneralBatchTableEnvironment;
+import org.apache.flink.table.api.BatchTableEnvironment;
 import org.apache.flink.table.api.Table;
 import org.apache.flink.table.api.TableEnvironment;
 import org.apache.flink.test.util.MultipleProgramsTestBase;
@@ -49,7 +49,7 @@ public class OrcTableSourceITCase extends MultipleProgramsTestBase {
 	public void testFullScan() throws Exception {
 
 		ExecutionEnvironment env = ExecutionEnvironment.getExecutionEnvironment();
-		GeneralBatchTableEnvironment tEnv = TableEnvironment.getTableEnvironment(env);
+		BatchTableEnvironment tEnv = TableEnvironment.getTableEnvironment(env);
 
 		OrcTableSource orc = OrcTableSource.builder()
 			.path(getPath(TEST_FILE_FLAT))
@@ -84,7 +84,7 @@ public class OrcTableSourceITCase extends MultipleProgramsTestBase {
 	public void testScanWithProjectionAndFilter() throws Exception {
 
 		ExecutionEnvironment env = ExecutionEnvironment.getExecutionEnvironment();
-		GeneralBatchTableEnvironment tEnv = TableEnvironment.getTableEnvironment(env);
+		BatchTableEnvironment tEnv = TableEnvironment.getTableEnvironment(env);
 
 		OrcTableSource orc = OrcTableSource.builder()
 			.path(getPath(TEST_FILE_FLAT))
