+++ flink-formats/flink-avro/src/test/java/org/apache/flink/table/runtime/batch/AvroTypesITCase.java
@@ -27,7 +27,7 @@ import org.apache.flink.formats.avro.generated.Fixed16;
 import org.apache.flink.formats.avro.generated.Fixed2;
 import org.apache.flink.formats.avro.generated.User;
 import org.apache.flink.formats.avro.utils.AvroKryoSerializerUtils;
-import org.apache.flink.table.api.GeneralBatchTableEnvironment;
+import org.apache.flink.table.api.BatchTableEnvironment;
 import org.apache.flink.table.api.Table;
 import org.apache.flink.table.api.TableEnvironment;
 import org.apache.flink.table.runtime.utils.TableProgramsClusterTestBase;
@@ -150,7 +150,7 @@ public class AvroTypesITCase extends TableProgramsClusterTestBase {
 		ExecutionEnvironment env = ExecutionEnvironment.getExecutionEnvironment();
 		env.getConfig().registerTypeWithKryoSerializer(LocalDate.class, AvroKryoSerializerUtils.JodaLocalDateSerializer.class);
 		env.getConfig().registerTypeWithKryoSerializer(LocalTime.class, AvroKryoSerializerUtils.JodaLocalTimeSerializer.class);
-		GeneralBatchTableEnvironment tEnv = TableEnvironment.getTableEnvironment(env, config());
+		BatchTableEnvironment tEnv = TableEnvironment.getTableEnvironment(env, config());
 
 		Table t = tEnv.fromDataSet(testData(env));
 		Table result = t.select("*");
@@ -177,7 +177,7 @@ public class AvroTypesITCase extends TableProgramsClusterTestBase {
 	@Test
 	public void testAvroStringAccess() throws Exception {
 		ExecutionEnvironment env = ExecutionEnvironment.getExecutionEnvironment();
-		GeneralBatchTableEnvironment tEnv = TableEnvironment.getTableEnvironment(env, config());
+		BatchTableEnvironment tEnv = TableEnvironment.getTableEnvironment(env, config());
 
 		Table t = tEnv.fromDataSet(testData(env));
 		Table result = t.select("name");
@@ -191,7 +191,7 @@ public class AvroTypesITCase extends TableProgramsClusterTestBase {
 	@Test
 	public void testAvroObjectAccess() throws Exception {
 		ExecutionEnvironment env = ExecutionEnvironment.getExecutionEnvironment();
-		GeneralBatchTableEnvironment tEnv = TableEnvironment.getTableEnvironment(env, config());
+		BatchTableEnvironment tEnv = TableEnvironment.getTableEnvironment(env, config());
 
 		Table t = tEnv.fromDataSet(testData(env));
 		Table result = t
@@ -206,7 +206,7 @@ public class AvroTypesITCase extends TableProgramsClusterTestBase {
 	@Test
 	public void testAvroToAvro() throws Exception {
 		ExecutionEnvironment env = ExecutionEnvironment.getExecutionEnvironment();
-		GeneralBatchTableEnvironment tEnv = TableEnvironment.getTableEnvironment(env, config());
+		BatchTableEnvironment tEnv = TableEnvironment.getTableEnvironment(env, config());
 
 		Table t = tEnv.fromDataSet(testData(env));
 		Table result = t.select("*");
