+++ flink-libraries/flink-table/src/main/java/org/apache/flink/table/api/java/package-info.java
@@ -19,9 +19,9 @@
 /**
  * <strong>Table API (Java)</strong><br>
  *
- * A {@link org.apache.flink.table.api.GeneralBatchTableEnvironment} can be used to create a
+ * A {@link org.apache.flink.table.api.BatchTableEnvironment} can be used to create a
  * {@link org.apache.flink.table.api.Table} from a {@link org.apache.flink.api.java.DataSet}.
- * Equivalently, a {@link org.apache.flink.table.api.GeneralStreamTableEnvironment} can be used to
+ * Equivalently, a {@link org.apache.flink.table.api.StreamTableEnvironment} can be used to
  * create a {@link org.apache.flink.table.api.Table} from a
  * {@link org.apache.flink.streaming.api.datastream.DataStream}.
  *
@@ -57,9 +57,9 @@
  * <p>
  * As seen above, a {@link org.apache.flink.table.api.Table} can be converted back to the
  * underlying API representation using
- * {@link org.apache.flink.table.api.GeneralBatchTableEnvironment#toDataSet(Table, java.lang.Class)},
- * {@link org.apache.flink.table.api.GeneralStreamTableEnvironment#toAppendStream(Table, java.lang.Class)}}, or
- * {@link org.apache.flink.table.api.GeneralStreamTableEnvironment#toRetractStream(Table, java.lang.Class)}}.
+ * {@link org.apache.flink.table.api.BatchTableEnvironment#toDataSet(Table, java.lang.Class)},
+ * {@link org.apache.flink.table.api.StreamTableEnvironment#toAppendStream(Table, java.lang.Class)}}, or
+ * {@link org.apache.flink.table.api.StreamTableEnvironment#toRetractStream(Table, java.lang.Class)}}.
  */
 
 package org.apache.flink.table.api.java;
