+++ flink-libraries/flink-table/src/main/scala/org/apache/flink/table/api/scala/DataSetConversions.scala
@@ -19,7 +19,7 @@ package org.apache.flink.table.api.scala
 
 import org.apache.flink.api.common.typeinfo.TypeInformation
 import org.apache.flink.api.scala._
-import org.apache.flink.table.api.{GeneralTableEnvironment, GeneralBatchTableEnvironment, Table, TableException}
+import org.apache.flink.table.api.{GeneralTableEnvironment, BatchTableEnvironment, Table, TableException}
 import org.apache.flink.table.expressions.Expression
 
 /**
@@ -47,11 +47,11 @@ class DataSetConversions[T](dataSet: DataSet[T], inputType: TypeInformation[T]) 
     * If not explicitly specified, field names are automatically extracted from the type of
     * the [[DataSet]].
     *
-    * @param tableEnv The [[GeneralBatchTableEnvironment]] in which the new [[Table]] is created.
+    * @param tableEnv The [[BatchTableEnvironment]] in which the new [[Table]] is created.
     * @param fields   The field names of the new [[Table]] (optional).
     * @return The resulting [[Table]].
     */
-  def toTable(tableEnv: GeneralBatchTableEnvironment, fields: Expression*): Table = {
+  def toTable(tableEnv: BatchTableEnvironment, fields: Expression*): Table = {
     if (fields.isEmpty) {
       tableEnv.fromDataSet(dataSet)
     } else {
