+++ flink-libraries/flink-table/src/main/scala/org/apache/flink/table/api/scala/DataStreamConversions.scala
@@ -18,7 +18,7 @@
 package org.apache.flink.table.api.scala
 
 import org.apache.flink.api.common.typeinfo.TypeInformation
-import org.apache.flink.table.api.{GeneralTableEnvironment, GeneralStreamTableEnvironment, Table, TableException}
+import org.apache.flink.table.api.{GeneralTableEnvironment, StreamTableEnvironment, Table, TableException}
 import org.apache.flink.table.expressions.Expression
 import org.apache.flink.streaming.api.scala.DataStream
 
@@ -47,11 +47,11 @@ class DataStreamConversions[T](dataStream: DataStream[T], inputType: TypeInforma
     * If not explicitly specified, field names are automatically extracted from the type of
     * the [[DataStream]].
     *
-    * @param tableEnv The [[GeneralStreamTableEnvironment]] in which the new [[Table]] is created.
+    * @param tableEnv The [[StreamTableEnvironment]] in which the new [[Table]] is created.
     * @param fields   The field names of the new [[Table]] (optional).
     * @return The resulting [[Table]].
     */
-  def toTable(tableEnv: GeneralStreamTableEnvironment, fields: Expression*): Table = {
+  def toTable(tableEnv: StreamTableEnvironment, fields: Expression*): Table = {
     if (fields.isEmpty) {
       tableEnv.fromDataStream(dataStream)
     } else {
